Up to Class 37
Forgot to commit yesterday, in this part of the course we're working on
the beginning of the frontend, just translating things from pure html to
JSX and doing some basic interactions.
There are a bunch of bugs though, and the guy who made the course
doesn't really pay much attention to the problems, so we're left on our
own to figure them out (which is good practice, but bad form when you've
paid for the course and expect the minimum support).